chore(stories): move Markdown stories under each viewer's storybook nav
Previously the Markdown stories lived in a separate top-level "Markdown"
group under packages/markdown/. That introduced a fourth top-level
category in Storybook's sidebar — inconsistent with the existing
PptxViewer / DocxViewer / XlsxViewer hierarchy. Move them to each
format package so they nest as <Viewer>/Markdown alongside Examples,
File Upload, and Debug JSON.

- packages/pptx/src/Markdown.stories.ts (title: PptxViewer/Markdown)
- packages/docx/src/Markdown.stories.ts (title: DocxViewer/Markdown)
- packages/xlsx/src/Markdown.stories.ts (title: XlsxViewer/Markdown)

Each story imports the conversion function from `@silurus/ooxml-markdown`
(added as workspace devDep on the three format packages so pnpm
symlinks resolve from inside each package's node_modules) and the
.wasm artifact via the format's own `./wasm/*.wasm?url` Vite import.

The placeholder `Markdown.stories.ts` and `vite-env.d.ts` under
packages/markdown/ are removed — that package is now back to a pure
library + CLI, no Storybook source of its own.
