+++ packages/xlsx/package.json
@@ -38,6 +38,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
+    "@silurus/ooxml-markdown": "workspace:*",
     "@types/pngjs": "^6.0.5",
     "pixelmatch": "^7.1.0",
     "playwright": "^1.59.1",
